finetune/requirements.txt
accelerate==1.6.0
annotated-types==0.7.0
blobfile==3.0.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
deepspeed==0.16.7
docker-pycreds==0.4.0
einops==0.8.1
filelock==3.13.1
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
hjson==3.1.0
huggingface-hub==0.31.2
idna==3.10
Jinja2==3.1.4
lxml==5.4.0
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.1.0
networkx==3.3
ninja==1.11.1.4
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
packaging==25.0
peft==0.15.2
pillow==11.0.0
platformdirs==4.3.8
protobuf==6.30.2
psutil==7.0.0
py-cpuinfo==9.0.0
pybind11==2.13.6
pycryptodomex==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
sentencepiece==0.2.0
sentry-sdk==2.28.0
setproctitle==1.3.6
six==1.17.0
smmap==5.0.2
sympy==1.13.3
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.4.0
torchaudio==2.4.0
torchvision==0.19.0
tqdm==4.67.1
transformers==4.50.0
triton==3.0.0
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.4.0
wandb==0.19.11
